# contract/openapi.yaml
openapi: 3.0.0
paths:
...
  /v1/clients:
    get:
      description: Requires clients:view. ?status filter.
      operationId: ClientsController_list_v1
      parameters:
        - name: status
          required: false
          in: query
          schema:
            default: active
            type: string
            enum:
              - active
              - inactive
              - all
      responses:
        "200":
          description: ""
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ClientResponse"
        "400": &a51
          description: Bad Request — validation failed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelopeDto"
        "401": &a52
          description: Unauthorized — missing/invalid token
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelopeDto"
        "403": &a53
          description: Forbidden — RBAC denied
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelopeDto"
        "404": &a54
          description: Not Found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelopeDto"
        "409": &a55
          description: Conflict — invalid state transition
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelopeDto"
        "422": &a56
          description: Unprocessable Entity — domain rule violated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelopeDto"
      security: &a57
        - bearer: []
      summary: List clients
      tags: &a58
        - Clients & Products
    post:
      description: Requires clients:create. Unique code → 409.
      operationId: ClientsController_create_v1
      parameters: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateClientDto"
      responses:
        "201":
          description: ""
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ClientResponse"
        "400": *a51
        "401": *a52
        "403": *a53
        "404": *a54
        "409": *a55
        "422": *a56
      security: *a57
      summary: Create a client
      tags: *a58
...
  /v1/clients/{id}/statements:
    post:
      description: Requires billing:create. Priced solely from client_billing_rates
        effective on each sale_date; no GST. Regenerating replaces the existing
        statement for the client+period (no duplicate).
      operationId: BillingGenerationController_generateStatement_v1
      parameters:
        - name: id
          required: true
          in: path
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/GenerateBillingDto"
      responses:
        "201":
          description: ""
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ClientStatementResponse"
        "400": &a131
          description: Bad Request — validation failed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelopeDto"
        "401": &a132
          description: Unauthorized — missing/invalid token
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelopeDto"
        "403": &a133
          description: Forbidden — RBAC denied
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelopeDto"
        "404": &a134
          description: Not Found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelopeDto"
        "409": &a135
          description: Conflict — invalid state transition
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelopeDto"
        "422": &a136
          description: Unprocessable Entity — domain rule violated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelopeDto"
      security: &a137
        - bearer: []
      summary: Generate a client statement (one line per customer)
      tags: &a138
        - Billing & Statements
  /v1/clients/{id}/invoices:
    post:
      description: Requires billing:create. total_commission = the client-billing
        statement total (billing stream only; never the rep commission payout).
        Regenerating replaces the existing invoice.
      operationId: BillingGenerationController_generateInvoice_v1
      parameters:
        - name: id
          required: true
          in: path
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/GenerateBillingDto"
      responses:
        "201":
          description: ""
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ClientInvoiceResponse"
        "400": *a131
        "401": *a132
        "403": *a133
        "404": *a134
        "409": *a135
        "422": *a136
      security: *a137
      summary: Generate a one-line commission invoice
      tags: *a138
  /v1/statements:
    get:
      description: Requires billing:view.
      operationId: StatementsController_list_v1
      parameters:
        - name: client_id
          required: false
          in: query
          description: Filter by client.
          schema:
            type: string
        - name: pay_period_id
          required: false
          in: query
          description: Filter by pay period.
          schema:
            type: string
      responses:
        "200":
          description: ""
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ClientStatementResponse"
        "400": &a139
          description: Bad Request — validation failed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelopeDto"
        "401": &a140
          description: Unauthorized — missing/invalid token
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelopeDto"
        "403": &a141
          description: Forbidden — RBAC denied
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelopeDto"
        "404": &a142
          description: Not Found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelopeDto"
        "409": &a143
          description: Conflict — invalid state transition
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelopeDto"
        "422": &a144
          description: Unprocessable Entity — domain rule violated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelopeDto"
      security: &a145
        - bearer: []
      summary: List generated statements
      tags: &a146
        - Billing & Statements
  /v1/statements/{id}:
    get:
      description: Requires billing:view.
      operationId: StatementsController_findOne_v1
      parameters:
        - name: id
          required: true
          in: path
          schema:
            type: string
      responses:
        "200":
          description: ""
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ClientStatementResponse"
        "400": *a139
        "401": *a140
        "403": *a141
        "404": *a142
        "409": *a143
        "422": *a144
      security: *a145
      summary: Get a statement with its lines
      tags: *a146
  /v1/statements/{id}/export:
    post:
      description: Requires billing:export. Updates the file_url reference; real
        render deferred.
      operationId: StatementsController_export_v1
      parameters:
        - name: id
          required: true
          in: path
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/BillingExportDto"
      responses:
        "201":
          description: ""
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BillingExportResultResponse"
        "400": *a139
        "401": *a140
        "403": *a141
        "404": *a142
        "409": *a143
        "422": *a144
      security: *a145
      summary: Export a statement (stub file)
      tags: *a146
  /v1/invoices:
    get:
      description: Requires billing:view.
      operationId: InvoicesController_list_v1
      parameters:
        - name: client_id
          required: false
          in: query
          description: Filter by client.
          schema:
            type: string
        - name: pay_period_id
          required: false
          in: query
          description: Filter by pay period.
          schema:
            type: string
      responses:
        "200":
          description: ""
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ClientInvoiceResponse"
        "400": &a147
          description: Bad Request — validation failed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelopeDto"
        "401": &a148
          description: Unauthorized — missing/invalid token
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelopeDto"
        "403": &a149
          description: Forbidden — RBAC denied
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelopeDto"
        "404": &a150
          description: Not Found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelopeDto"
        "409": &a151
          description: Conflict — invalid state transition
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelopeDto"
        "422": &a152
          description: Unprocessable Entity — domain rule violated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelopeDto"
      security: &a153
        - bearer: []
      summary: List generated invoices
      tags: &a154
        - Billing & Statements
  /v1/invoices/{id}:
    get:
      description: Requires billing:view.
      operationId: InvoicesController_findOne_v1
      parameters:
        - name: id
          required: true
          in: path
          schema:
            type: string
      responses:
        "200":
          description: ""
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ClientInvoiceResponse"
        "400": *a147
        "401": *a148
        "403": *a149
        "404": *a150
        "409": *a151
        "422": *a152
      security: *a153
      summary: Get an invoice
      tags: *a154
  /v1/invoices/{id}/export:
    post:
      description: Requires billing:export. Updates the file_url reference; real
        render deferred.
      operationId: InvoicesController_export_v1
      parameters:
        - name: id
          required: true
          in: path
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/BillingExportDto"
      responses:
        "201":
          description: ""
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BillingExportResultResponse"
        "400": *a147
        "401": *a148
        "403": *a149
        "404": *a150
        "409": *a151
        "422": *a152
      security: *a153
      summary: Export an invoice (stub file)
      tags: *a154
...
info:
  title: Redwave ERP / HRM API
  description: Versioned REST contract for the Redwave platform. Each protected
    endpoint notes the required (module, action) permission enforced server-side
    by the RBAC guard.
  version: "1.0"
  contact: {}
...
servers: []
components:
  securitySchemes:
    bearer:
      scheme: bearer
      bearerFormat: JWT
      type: http
  schemas:
    ErrorBodyDto:
      type: object
      properties:
        code:
          type: string
          example: UNPROCESSABLE_ENTITY
          description: Stable machine code for the error.
        message:
          type: string
          example: tier brackets must be contiguous (each min_count = previous max_count +
            1)
        details:
          type: object
          description: Optional structured detail (e.g. validation messages, an
            unpriced-products list).
          additionalProperties: true
      required:
        - code
        - message
    ErrorEnvelopeDto:
      type: object
      properties:
        error:
          $ref: "#/components/schemas/ErrorBodyDto"
      required:
        - error
...
    ClientResponse:
      type: object
      properties:
        id:
          type: string
        client_code:
          type: string
          example: VF
        name:
          type: string
        market:
          type: string
          enum:
            - CA
            - US
        supplies_mpu_id:
          type: boolean
        is_active:
          type: boolean
        created_at:
          type: string
          format: date-time
      required:
        - id
        - client_code
        - name
        - market
        - supplies_mpu_id
        - is_active
        - created_at
    CreateClientDto:
      type: object
      properties:
        client_code:
          type: string
          example: VF
          description: Unique client code (e.g. VF / RF / CTI).
        name:
          type: string
          example: Valley Fiber
        market:
          type: string
          enum:
            - CA
            - US
        supplies_mpu_id:
          type: boolean
          description: Whether the partner supplies per-house MPU IDs.
      required:
        - client_code
        - name
        - market
        - supplies_mpu_id
...
    GenerateBillingDto:
      type: object
      properties:
        pay_period_id:
          type: string
          description: "The pay period to bill (governs which sales by sale_date, #7)."
      required:
        - pay_period_id
    ClientStatementLineResponse:
      type: object
      properties:
        id:
          type: string
        statement_id:
          type: string
        sale_id:
          type: string
        customer_name:
          type: string
        products_summary:
          type: string
          example: Internet, TV, Home Phone
        line_total:
          type: string
          example: "90.00"
          description: Decimal string. Server-priced line total.
      required:
        - id
        - statement_id
        - sale_id
        - customer_name
        - products_summary
        - line_total
    ClientStatementResponse:
      type: object
      properties:
        id:
          type: string
        client_id:
          type: string
        pay_period_id:
          type: string
        total_amount:
          type: string
          example: "140.00"
          description: Decimal string. Server-computed statement total (no GST).
        file_url:
          type: string
        generated_by:
          type: string
        generated_at:
          type: string
          format: date-time
        lines:
          description: Present on generate/detail; absent on the list.
          type: array
          items:
            $ref: "#/components/schemas/ClientStatementLineResponse"
      required:
        - id
        - client_id
        - pay_period_id
        - total_amount
        - file_url
        - generated_by
        - generated_at
    ClientInvoiceResponse:
      type: object
      properties:
        id:
          type: string
        client_id:
          type: string
        pay_period_id:
          type: string
        total_commission:
          type: string
          example: "140.00"
          description: Decimal string = the billing-stream statement total (#3).
        file_url:
          type: string
        generated_at:
          type: string
          format: date-time
      required:
        - id
        - client_id
        - pay_period_id
        - total_commission
        - file_url
        - generated_at
    BillingExportDto:
      type: object
      properties:
        format:
          type: string
          enum:
            - pdf
            - excel
          example: excel
      required:
        - format
    BillingExportResultResponse:
      type: object
      properties:
        statement_id:
          type: string
          description: Set when exporting a statement.
        invoice_id:
          type: string
          description: Set when exporting an invoice.
        format:
          type: string
          enum:
            - pdf
            - excel
        file_url:
          type: string
        content:
          type: string
          description: The serialized row payload (real render deferred).
      required:
        - format
        - file_url
        - content
...